requirements.txt
google_api_python_client==2.129.0
gspread==6.1.2
numpy==1.26.4
protobuf==5.27.0
pytz==2024.1
streamlit==1.34.0
streamlit_option_menu==0.3.12
